+++ src/vg_nav/cfg/ThresholdConfig.cfg
@@ -1,5 +1,5 @@
 #!usr/bin/env python3
-PACKAGE = "vg_gpn"
+PACKAGE = "vg_nav"
 
 from dynamic_reconfigure.parameter_generator_catkin import *
 gen = ParameterGenerator()
@@ -18,4 +18,4 @@ gen.add("max_threshold_blue",  double_t, 0, "max Threshold for the blue channel"
 
 
 
-exit(gen.generate(PACKAGE, "vg_gpn", "Threshold"))
+exit(gen.generate(PACKAGE, "vg_nav", "Threshold"))
