+++ src/vg_nav/config/vg_nav.yaml
@@ -4,24 +4,22 @@ oc_srfc_rds : 8         # radius of the occupancy surface
 org_oc_srfc_rds_viz: 9 # visualization radius of the occupancy surface in rviz
 velodyne_frame: "velodyne"
 scan_frame: "velodyne"
+log_data: False
+sim: False
+gmtry_nav_mode: True # True means geometry, False means visual-geometry
+v_nvgblti_th: 16.5 # should not be int # any lnp has less than 20 vis-navigable rows is then non-navigable
+
+######## Goal #######
+gl_x: 12  # x-coordinate of the goal  
+gl_y: 6 # y-coordinate of the goal 
+
 
 ######## Nav #########
 pcl_skp: 3  #Downsampling pointcloud 
-
-gl_x: 12  # x-coordinate of the goal 
-# gl_y: -11 # 10 # y-coordinate of the goal 
-gl_y: 6 # y-coordinate of the goal 
-
 gp_nav_indpts_sz   : 400 # NO. of inducing points 
 var_km : 0.5
 var_kv : 0.03
-
-
-
 sb_gl_dst     : 5
-
-gmtry_nav_mode: 0
-v_nvgblti_th: 16.5 # should not be int # any lnp has less than 20 vis-navigable rows is then non-navigable
 k_dir       : 0.0   # direction factor
 k_dst       : 0.99 #0.3  # distance factor  
 k_nav       : 0.5 #0.7
@@ -57,20 +55,6 @@ angular_tolerance : 0.1     #
 
 max_linear_acceleration : 0.1  #
 max_angular_acceleration: 0.1   #
-
-
-# max_linear_speed  : 1.0   #
-# min_linear_speed  : -1.0   #
-
-# max_angular_speed : 1.2     #
-# min_angular_speed : -1.2    #
-
-# linear_tolerance  : 0.4        #
-# angular_tolerance : 0.4     #
-
-# max_linear_acceleration : 0.1  #
-# max_angular_acceleration: 0.1   #
-
 
 # v= kp* dst 
 # w= ka*a + kb*b
